Enhanced WebSocket server with authentication credential support: - Updated TerminalId type from int to long across files. - Added new request/response types for auth credentials. - Introduced `SetAuthCredentialsReq` and `SetAuthCredentialsReqResponse` classes. - Updated WebSocket address and terminal ID constants in `Program.cs`.

+++ SimpleWebSocketServer.SIBS.Front/Models/NewTerminalConnectedReq.cs
@@ -11,7 +11,7 @@ namespace SimpleWebSocketServer.SIBS.Front.Models
 
 
         [JsonProperty("terminalId")]
-        public int TerminalId { get; set; }
+        public long TerminalId { get; set; }
 
         [JsonProperty("type")]
         [JsonConverter(typeof(StringEnumConverter))]
